have spec specific to project representer rendering
